Add transitions for cards and a detail view
